Remove the carbon emission lines from the Mappings dust continuum (#15)
Adjust the function that creates the stored table for the Mappings SED family templates so that it removes the three overly prominent Carbon emission lines from the dust continuum. The problem is that these emission lines are defined in the templates through just a single point in a coarse wavelength grid. As a result they are much wider than they should be and thus they carry much more energy than physically justified.