Added script to convert from raw HEALPix data to a projected image. (#17)
Description
This pull request includes a file that contains the necessary functions to convert the raw HEALPixSkyInstrument output to a projected image, and a script that uses these functions to convert a raw HEALPixSkyInstrument FITS image to a flat FITS image using an appropriate projection.

Motivation
Without this functionality, the output of the HEALPixSkyInstrument is practically useless, since the positions of the pixels in the raw FITS file are incorrect when analysing the FITS file with a standard FITS viewer or script. This pull request provides the very basic functionality required to convert HEALPixSkyInstrument to images that are very similar to those produced by the AllSkyInstrument. This functionality can later be extended to produce FITS projections with a different centre (so that one HEALPixSkyInstrument can mimic multiple AllSkyInstruments at the same position), and to include the functionality required to perform an analysis on the celestial sphere.