Healpix improvements (#18)
This commit contains a thorough refactoring of the HEALPix support. A HEALPix grid is now represented internally as an object that caches some information for more efficient access. This object has functionality to access the pixels in various ways, including the option to retrieve all pixels within an annulus around a pixel. The existing HEALPix projection script has been extended with an option to rotate the sky to an arbitrary new cross hair, and has been updated so that the total flux is now conserved, independent of the chosen resolution. A new conversion script has been added to convert the SKIRT HEALPix output into a FITS format that is compatible with the official HEALPix software and tools as used by e.g. Planck.